20: header responsivity -updated folder and file structure -initialized mobile component
