+++ app/components/menu-bar/mobile/mobile.component.tsx
@@ -1,0 +1,12 @@
+import { FC } from "react";
+import StyledWrapper from "./mobile.style";
+
+const Mobile: FC = () => {
+  return (
+    <StyledWrapper container className="menu-bar-mobile--grid-container">
+      mobile
+    </StyledWrapper>
+  );
+};
+
+export default Mobile;
