+++ app/components/menu-bar/mobile/mobile.style.tsx
@@ -1,0 +1,8 @@
+"use client";
+
+import { Grid } from "@mui/material";
+import { styled } from "@mui/material/styles";
+
+const StyledWrapper = styled(Grid)(() => ({}));
+
+export default StyledWrapper;
